requirements.txt
Pillow==8.2.0
plotly==5.4.0
pandas==1.4.2
torch==1.9.0
torchvision==0.10.0
timm==0.4.12
captum==0.5.0
watchdog
streamlit==1.16.0
streamlit-drawable-canvas==0.9.0
scipy>=1.0.1
scikit-image==0.17.2
